% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{accuracy_metric}
\alias{accuracy_metric}
\title{Accuracy metrics}
\usage{
accuracy_metric(y_var_true, y_var_pred, metric = "mape")
}
\arguments{
\item{y_var_true}{Numeric. Observed value.}

\item{y_var_pred}{Numeric. Predicted value.}

\item{metric}{String. Name of the accuracy metric.}
}
\value{
Numeric.
}
\description{
Mean Average Prediction Error forecast accuracy metric
}
\examples{
\dontrun{
accuracy_metric()
}
}
