% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansing.R
\name{clean_ts}
\alias{clean_ts}
\title{Time Series Cleansing}
\usage{
clean_ts(
  .data,
  y_var,
  date_var,
  method = "winsorize",
  freq = 12,
  na_exclude = NULL,
  replace_y_var = TRUE
)
}
\arguments{
\item{.data}{DataFrame or tibble.}

\item{y_var}{String. Colname of the variable to be cleansed.}

\item{date_var}{String. Colname of time serie index.}

\item{method}{Method to be applied to the time series. Options: kalman, winsorize, 
mean, median, nearest observation, interpolation. If more than one method is chosen, a set of new columns
is automatically generated.}

\item{freq}{Time series frequency.}

\item{na_exclude}{Vector of names that **should not** be used to replace the observation by NA.}

\item{replace_y_var}{Logical. If replace = TRUE (default) it will 
replace the previous y_var by its cleansed/imputated version. Otherwise, a "y_var_clean" column will
be attached to the data matrix.}
}
\value{

}
\description{
This function collects the data and applies imputation, time series cuts and leading zeros filtering
}
