% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansing.R
\name{cleansing}
\alias{cleansing}
\title{Cleansing function}
\usage{
cleansing(data, method, frequency = 12, regressors = NULL, keep_old = FALSE)
}
\arguments{
\item{data}{tibble or data frames}

\item{method}{string. Method to be applied to the time series. Options: kalman, winsorize, 
mean, median, nearest, interpolation}

\item{frequency}{time series frequency}

\item{regressors}{string. vector of regressors used to mark as NA's to apply cleansing methods.}

\item{keep_old}{logical. Keep or not the old dependent variable (time series)}
}
\value{

}
\description{
This function collects the data and applies imputation, time series cuts and leading zeros filtering
}
