% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_funs.R
\name{detect_seasonality}
\alias{detect_seasonality}
\title{Detect seasonality}
\usage{
detect_seasonality(data)
}
\arguments{
\item{data}{numeric. Time series data}
}
\value{

}
\description{
This is a consensus tool to decide if a time series exhibit regular patterns over time.
}
\details{
The consensus is formed as Kruskal test, seasonal strength, and statistical significance
of a linear regression model. Rule: if 2 out of 3 methods show evidence in favor of seasonality
then, the data will employ seasonal models to generate the forecast.
}
