% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_arima.R
\name{fit_arima}
\alias{fit_arima}
\title{Fit Auto Regressive Integrated Moving Average Model}
\usage{
fit_arima(.data, parameter = NULL)
}
\arguments{
\item{.data}{Data frame or tibble with a response variable.}

\item{parameter}{List. Combination of parameter to estimate the model.}
}
\value{
data-frame
}
\description{
Fit Auto Regressive Integrated Moving Average Model
}
\examples{
\dontrun{
get_arima()
}
}
