% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_croston.R
\name{fit_croston}
\alias{fit_croston}
\title{Fit Croston model}
\usage{
fit_croston(.data, parameter = NULL)
}
\arguments{
\item{.data}{Data frame or tibble with a response variable.}

\item{parameter}{list: Combination of parameter to estimate the model.}
}
\value{
data-frame
}
\description{
Fit Croston model
}
\examples{
\dontrun{
get_croston()
}
}
