% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit_gam}
\alias{fit_gam}
\title{Fit Generalized Additive Model}
\usage{
fit_gam(.data, parameter = list())
}
\arguments{
\item{.data}{tibble/data.frame:}

\item{parameter}{list: hyperparameter and parameter to be applied.}
}
\value{

}
\description{
Fit Generalized Additive Model
}
