% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit a Generalized Linear Model}
\usage{
fit_glm(.data, parameter)
}
\arguments{
\item{.data}{tibble/data.frame: matrix of response and covariates.}

\item{parameter}{list: Combination of parameter to estimate the model.}
}
\value{
data-frame
}
\description{
Fit a Generalized Linear Model
}
\examples{
\dontrun{
get_glm()
}
}
