% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glmnet.R
\name{fit_glmnet}
\alias{fit_glmnet}
\title{Fit a Regularized Generalize Linear Model}
\usage{
fit_glmnet(.data, parameter)
}
\arguments{
\item{.data}{tibble/data.frame: matrix of response and covariates.}

\item{parameter}{list: Combination of parameter to estimate the model.}
}
\value{
data-frame
}
\description{
Fit a Regularized Generalize Linear Model
}
\examples{
\dontrun{
get_glmnet()
}
}
\author{
Obryan Poyser
}
