% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{fit_ts}
\alias{fit_ts}
\title{Fitting models for time-series data}
\usage{
fit_ts(.data, y_var, date_var, model, parameter = NULL)
}
\arguments{
\item{.data}{data-frame or tibble}

\item{y_var}{Column name of the variable to be forecasted}

\item{date_var}{Column name of the time index}

\item{model}{String. Name of the model to be estimated.}

\item{parameter}{List. Optional set of parameters to estimate models.}
}
\value{
data-frame
}
\description{
Wrapper of single function call to simplify the estimation
}
\examples{
\dontrun{
fit_ts()
}
}
