% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_ts.R
\name{forecast_ts}
\alias{forecast_ts}
\title{Generate forecast}
\usage{
forecast_ts(.fit, new_data = NULL, horizon = 12)
}
\arguments{
\item{.fit}{fit object}

\item{new_data}{tibble}

\item{horizon}{numeric}
}
\value{

}
\description{
Generate forecast
}
