% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arima.R
\name{get_arima}
\alias{get_arima}
\title{Fit Auto Regressive Integrated Moving Average Model}
\usage{
get_arima(.data, y_var, is_seasonal = TRUE, parameter = NULL, freq)
}
\arguments{
\item{.data}{Data frame or tibble with a response variable.}

\item{y_var}{String. Column name of the time series to be forecasted.}

\item{is_seasonal}{Logical. Defines if the optimization takes into account seasonal parameters.}

\item{parameter}{List. Combination of parameter to estimate the model.}

\item{freq}{Numeric. Time series frequency.}
}
\value{

}
\description{
Fit Auto Regressive Integrated Moving Average Model
}
