% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_croston.R
\name{get_croston}
\alias{get_croston}
\title{Fit Croston model}
\usage{
get_croston(.data, y_var, horizon = 10, parameter = NULL)
}
\arguments{
\item{.data}{Data frame or tibble with a response variable.}

\item{y_var}{String. Column name of the time series to be forecasted.}

\item{horizon}{Numeric. Number of periods to forecast.}

\item{parameter}{List. Parameter to be used for estimation, looks for alpha parameter.}
}
\value{

}
\description{
Fit Croston model
}
