% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_optimization.R
\name{get_default_optim_conf}
\alias{get_default_optim_conf}
\title{Default optimization configuration}
\usage{
get_default_optim_conf(test_size = NULL, lag = NULL, optim_profile = NULL)
}
\arguments{
\item{test_size}{integer: number of periods to be considered to test data.}

\item{lag}{integer: number of periods ahead to predict.}

\item{optim_profile}{string: name of the optimization profile {light, medium, complete}}
}
\value{

}
\description{
Default optimization configuration
}
