% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_optimization.R
\name{get_default_pipeline_conf}
\alias{get_default_pipeline_conf}
\title{Default pipeline configuration}
\usage{
get_default_pipeline_conf()
}
\value{

}
\description{
Default pipeline configuration
}
