% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{get_design_matrix}
\alias{get_design_matrix}
\title{Generate trend and seasonal components for regression based models}
\usage{
get_design_matrix(
  .data,
  date_var = NULL,
  freq = NULL,
  parameter = NULL,
  to_dummy = TRUE
)
}
\arguments{
\item{.data}{DataFrame or tibble}

\item{date_var}{String. Column name of the time index variable}

\item{freq}{Numeric. Time series frequency}

\item{parameter}{List.}

\item{to_dummy}{Logical. Convert design matrix factors to binary.}
}
\value{

}
\description{
Generate trend and seasonal components for regression based models
}
