% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dyn_theta.R
\name{get_dyn_theta}
\alias{get_dyn_theta}
\title{Fit Dynamic Theta Model}
\usage{
get_dyn_theta(.data, y_var, horizon = 12, parameter = NULL)
}
\arguments{
\item{.data}{Data frame or tibble.}

\item{y_var}{String. Column name of the time series to be forecasted.}

\item{horizon}{Numeric. Number of periods to forecast.}

\item{parameter}{List.}
}
\value{
data-frame
}
\description{
Fit Dynamic Theta Model
}
\examples{
\dontrun{
get_ets()
}
}
