% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Generate forecast figures}
\usage{
get_forecast(.fit_output, x_data = NULL, horizon = NULL, tune = FALSE)
}
\arguments{
\item{.fit_output}{Class ".fit_output".}

\item{x_data}{Data frame. Design matrix to calculate predicted/forecast figures.}

\item{horizon}{Numeric. Number of periods ahead to forecast.}

\item{tune}{Logical. Results will be used to define the best model.}
}
\value{

}
\description{
Generate forecast figures
}
