% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glm.R
\name{get_glm}
\alias{get_glm}
\title{Fit a Generalized Linear Model}
\usage{
get_glm(.data, y_var, date_var, parameter)
}
\arguments{
\item{.data}{Data frame or tibble with a response variable.}

\item{y_var}{String. Column name of the time series to be forecasted.}

\item{date_var}{String.Column name of the time series to be forecasted.}

\item{parameter}{List. Trend discount and time weight parameters.}
}
\value{
data-frame
}
\description{
Fit a Generalized Linear Model
}
\examples{
\dontrun{
get_glm()
}
}
