% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glmnet.R
\name{get_glmnet}
\alias{get_glmnet}
\title{Fit a Regulalized Generalized Linear Model}
\usage{
get_glmnet(data, hyperparam, config)
}
\arguments{
\item{data}{Data-frame. with response and regressor variables.}

\item{hyperparam}{List. Hyperparameters to be used for estimation. There are 4 hyperparameters: first, 
*alpha* in the space [0,1] controls whether it is a Ridge (L2) a LASSO (L1) shrinkage method, respectively.
Any number that lies between is considered as ElasticNet regression, a combination of both regularizations.
The other 2 hyperparameters are time weights and trend discount.}

\item{config}{List. Configuration that defines the task to be performed.}
}
\value{

}
\description{
Fit a Regulalized Generalized Linear Model
}
\examples{
\dontrun{
 get_glmnet()
}
}
\author{
Obryan Poyser
}
