% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{get_graph_stat}
\alias{get_graph_stat}
\title{Generate graphics for time series insights}
\usage{
get_graph_stat(
  insight_data,
  graph_type,
  conf = list(min_limit_zero = F, gam_forecast = F)
)
}
\arguments{
\item{insight_data}{tibble.}

\item{graph_type}{string: {derivative, forecast, marginal}.}

\item{conf}{list: graph options.}
}
\value{

}
\description{
Generate graphics for time series insights
}
