% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_optimization.R
\name{get_hyperpar_sample}
\alias{get_hyperpar_sample}
\title{Hyperparameter sample generator}
\usage{
get_hyperpar_sample(ts_model, parameter)
}
\arguments{
\item{ts_model}{string}

\item{parameter}{list: initial set of hyperparameters.}
}
\value{

}
\description{
Hyperparameter sample generator
}
