% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{get_insight_data}
\alias{get_insight_data}
\title{Get GAM features}
\usage{
get_insight_data(oc_data, key, parameter)
}
\arguments{
\item{oc_data}{tibble}

\item{key}{string: name of the key to be processed.}

\item{parameter}{list: parameters and hyperparameters.}
}
\value{

}
\description{
Get GAM features
}
