% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{get_oc_data}
\alias{get_oc_data}
\title{Get Operational Cycle data}
\usage{
get_oc_data(db = list(), countries, gbus, date_cycle)
}
\arguments{
\item{db}{list: list of DB from the OC.}

\item{countries}{string: vector of country codes.}

\item{gbus}{string: GBU to be extracted.}

\item{date_cycle}{date: date of the cycle in yyyy-mm-dd format.}
}
\value{

}
\description{
Get Operational Cycle data
}
