% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prophet.R
\name{get_prophet}
\alias{get_prophet}
\title{Fit Prophet Model}
\usage{
get_prophet(.data, y_var, horizon = 12, parameter = NULL)
}
\arguments{
\item{.data}{Data frame or tibble.}

\item{y_var}{String. Column name of the time series to be forecasted.}

\item{horizon}{Numeric. Number of periods to forecast.}

\item{parameter}{List.}
}
\value{
data-frame
}
\description{
Fit Prophet Model
}
\examples{
\dontrun{
get_ets()
}
}
