% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_funs.R
\name{get_regressors}
\alias{get_regressors}
\title{Generate trend and seasonal components for regression based models}
\usage{
get_regressors(.data, date_var = NULL, frequency = NULL, bind = TRUE)
}
\arguments{
\item{.data}{DataFrame or tibble}

\item{date_var}{String. Column name of the time index variable}

\item{frequency}{Numeric. Time series frequency}

\item{bind}{Logical. Whether or not to bind the regressors to the original data. Default = TRUE.}
}
\value{

}
\description{
Generate trend and seasonal components for regression based models
}
