% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{get_seas_me}
\alias{get_seas_me}
\title{Calculate Marginal Seasonal Effects}
\usage{
get_seas_me(
  .data,
  min_date = NULL,
  max_date = NULL,
  robust_lm = F,
  ci_me = 0.95
)
}
\arguments{
\item{.data}{tibble/data.frame}

\item{min_date}{date: set mininum date.}

\item{max_date}{date: set maximum date.}

\item{robust_lm}{logical: whether or not to use robust regression.}

\item{ci_me}{numeric: confidence interval for marginal effects.}
}
\value{

}
\description{
Calculate Marginal Seasonal Effects
}
