% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snaive.R
\name{get_seasonal_naive}
\alias{get_seasonal_naive}
\title{Fit a Seasonal Naive model}
\usage{
get_seasonal_naive(.data, y_var, horizon)
}
\arguments{
\item{.data}{Data frame or tibble with a response variable.}

\item{y_var}{String. Column name of the time series to be forecasted.}

\item{horizon}{Numeric. Number of periods ahead to forecast.}
}
\value{

}
\description{
Fit a Seasonal Naive model
}
