% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_svm.R
\name{get_svm}
\alias{get_svm}
\title{Fit SVM model}
\usage{
get_svm(.data, y_var, parameter = NULL)
}
\arguments{
\item{.data}{Data frame or tibble.}

\item{y_var}{String. Column name of the time series to be forecasted.}

\item{parameter}{List. Combination of parameter to estimate the model.}
}
\value{
dataframe
}
\description{
Fit SVM model
}
\examples{
\dontrun{
get_ets()
}
}
