% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{get_tables}
\alias{get_tables}
\title{Generate dashboard tables}
\usage{
get_tables(insight_data, table_type)
}
\arguments{
\item{insight_data}{tibble}

\item{table_type}{character}
}
\value{

}
\description{
Generate dashboard tables
}
\examples{
\dontrun{
get_tables()
}
}
