% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{get_time_weights}
\alias{get_time_weights}
\title{Generate time weights}
\usage{
get_time_weights(y_var, time_weight)
}
\arguments{
\item{y_var}{numeric. Time series vector data.}

\item{time_weight}{Numeric. How rapidly recent observations weight for estimation.}
}
\value{
Numerical vector.
}
\description{
Method to define data points' weights for regression based models.
}
\examples{
\dontrun{
get_time_weights()
}
}
\author{
Obryan Poyser
}
