% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{get_trend_decay}
\alias{get_trend_decay}
\title{Generate trend discounts}
\usage{
get_trend_decay(y_var_length, trend_decay, horizon = NULL, lag = NULL)
}
\arguments{
\item{y_var_length}{integer: Length of the time series}

\item{trend_decay}{numeric: How rapidly the trend reach the stability.}

\item{horizon}{integer: How far in time to produce trend discounts.}

\item{lag}{integer: Lag to be used for cross-validation purposes.}
}
\value{
Numerical vector.
}
\description{
Regression based models use a linear trend to account for the change in level over time. 
In practical terms, it is measured as a vector of equidistant integers.
Often, the trend component can significantly impact the forecast in the long run. One way to solve
this issue is to apply a "discount" on the trend vector, henceforth, reducing the marginal effect on the 
predictions.
}
\examples{
\dontrun{
get_trend_discounts()
}
}
\author{
Obryan Poyser
}
