% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tslm.R
\name{get_tslm}
\alias{get_tslm}
\title{Fit Simple Time Series Linear Model}
\usage{
get_tslm(.data, y_var, parameter = NULL)
}
\arguments{
\item{.data}{Data frame or tibble.}

\item{y_var}{String. Column name of the time series to be forecasted.}

\item{parameter}{List.}
}
\value{
data-frame
}
\description{
Fit Simple Time Series Linear Model
}
\examples{
\dontrun{
get_ets()
}
}
