% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_data}
\alias{import_data}
\title{Importing data from local and shared sources}
\usage{
import_data(source_conf)
}
\arguments{
\item{source_conf}{list: configuration for data source.}
}
\value{

}
\description{
Importing data from local and shared sources
}
