% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleansing.R
\name{impute_ts}
\alias{impute_ts}
\title{Time series imputation}
\usage{
impute_ts(.data, freq = numeric(), imputation_config = list(), ...)
}
\arguments{
\item{.data}{tibble/data.frame: data.}

\item{freq}{integer: frequency of the data.}

\item{imputation_config}{list: options for imputation methods. The names can take the 
following configurations: impute_method = {kalman, median, mean, nearest, interpolation}
add_transformation = logical, na_regressor = logical (replaces response variable with NA's and imputates)
, na_missing_dates = logical (replaces response variable with NA's and imputates when missing dates)
, na_value = character (replace NA's by a specific value)}

\item{...}{Other parameters,
.}
}
\value{
data-frame
}
\description{
Time series imputation
}
\examples{
\dontrun{
impute_ts()
}
}
\author{
Obryan Poyser
}
