% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{lag_features}
\alias{lag_features}
\title{Autoregressive or lags features}
\usage{
lag_features(.data, lag_var = list())
}
\arguments{
\item{.data}{data-frame or tibble}

\item{lag_var}{list: defines the number of lag periods for a given variable, for instance, 
list(y_var = 1) will generate a lag (autoregressive) variable of 1 period.}
}
\value{

}
\description{
This function generate AR for response and features
}
