% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{log_update}
\alias{log_update}
\title{Logger updater}
\usage{
log_update(module = character(), key = character(), new_log = list())
}
\arguments{
\item{module}{string: name of the module.}

\item{key}{string}

\item{new_log}{list: log to be added.}
}
\value{

}
\description{
Logger updater
}
