% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{long_to_wide}
\alias{long_to_wide}
\title{Long to wide regressor column}
\usage{
long_to_wide(.data)
}
\arguments{
\item{.data}{tibble/data.frame}
}
\value{

}
\description{
Long to wide regressor column
}
