% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{ma_features}
\alias{ma_features}
\title{Moving Average feature engineering}
\usage{
ma_features(.data, ma_var = list())
}
\arguments{
\item{.data}{tibble}

\item{ma_var}{list: list of strings defining the desired configuration.}
}
\value{

}
\description{
Moving Average feature engineering
}
