% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{make_reg_matrix}
\alias{make_reg_matrix}
\title{Generate regression matrix for regression based models}
\usage{
make_reg_matrix(.fit_output, x_data = NULL, horizon = NULL)
}
\arguments{
\item{.fit_output}{Inherited data and meta-data from a model fit.}

\item{x_data}{Optional DataFrame to use as design matrix.}

\item{horizon}{Numeric. How far in time to produce a synthetic design matrix.}
}
\value{
data-frame or tibble
}
\description{
This function takes the metadata from the fit to generate a template of the
dependent variables to make forecast.
}
\examples{
\dontrun{
make_reg_matrix()
}
}
