% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{mod_stat_data}
\alias{mod_stat_data}
\title{Internal modification of statistical data}
\usage{
mod_stat_data(insight_data, date = NULL)
}
\arguments{
\item{insight_data}{tibble/data.frame}

\item{date}{date: set minimum date.}
}
\value{

}
\description{
Internal modification of statistical data
}
