% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansing.R
\name{na_winsorize}
\alias{na_winsorize}
\title{Winsorize imputation}
\usage{
na_winsorize(y_var, na_marker = NULL, freq = 12, include = FALSE)
}
\arguments{
\item{y_var}{Numeric. Time series to be cleansed.}

\item{na_marker}{logical or binary indicator that defines which observation should be take out of the
loess decomposition.}

\item{freq}{Numeric. Time series frequency, 12 by default.}

\item{include}{Logical. Whether to print all time series components.}
}
\value{
numeric cleansed series
}
\description{
This function replaces outliers by an expected value given the 5% and 95% percentiles of the
error component plus the denoised series.
}
\examples{
\dontrun{
na_winsorize(AirPassengers)
}
}
\author{
Obryan Poyser
}
