% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_glmnet.R
\name{optim_glmnet}
\alias{optim_glmnet}
\title{Hyperparameter optimization for a glmnet model}
\usage{
optim_glmnet(.data, hyperparam, random_sample = 0.2, config)
}
\arguments{
\item{hyperparam}{List. Hyperparameters to be used for estimation. There are 4 hyperparameters: first, 
*alpha* in the space [0,1] controls whether it is a Ridge (L2) a LASSO (L1) shrinkage method, respectively.
Any number that lies between is considered as ElasticNet regression, a combination of both regularizations.
The other 2 hyperparameters are time weights and trend discount.}

\item{random_sample}{Numeric. In the space [0,1] defines the Random Search to Grid Search hyperparameter
search process, respectively. It is set as 20% random sample by default.}

\item{config}{List. Configuration that defines the task to be performed.}

\item{data}{Data-frame.}
}
\value{

}
\description{
Hyperparameter optimization for a glmnet model
}
