% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_optimization.R
\name{optim_int}
\alias{optim_int}
\title{Time series optimization internal function}
\usage{
optim_int(
  .data,
  ts_model = character(),
  optim_conf = list(),
  parameter = list(),
  export_fit = FALSE
)
}
\arguments{
\item{.data}{tibble/data.frame}

\item{ts_model}{string: model to be optimized.}

\item{optim_conf}{list: optimization configuration.}

\item{parameter}{list: parameters and hyperparameters.}

\item{export_fit}{logical: whether or not to include fit object.}
}
\value{

}
\description{
Time series optimization internal function
}
