% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_optimization.R
\name{optim_ts}
\alias{optim_ts}
\title{High order optimization function}
\usage{
optim_ts(
  .data,
  ts_model = character(),
  optim_conf = list(),
  parameter = list(),
  export_fit = FALSE
)
}
\arguments{
\item{.data}{tibble}

\item{ts_model}{character}

\item{optim_conf}{list of optimization configuration}

\item{parameter}{list of parameters}

\item{export_fit}{logical}
}
\value{

}
\description{
High order optimization function
}
