% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_optimization.R
\name{pipeline_ts}
\alias{pipeline_ts}
\title{Execute pipeline and return optimization results}
\usage{
pipeline_ts(.data_presc, .pipeline_conf = list())
}
\arguments{
\item{.data_presc}{tibble}

\item{pipeline_conf}{list: configuration of pipelines' modules}
}
\value{

}
\description{
Execute pipeline and return optimization results
}
