% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot time series forecast}
\usage{
plot_ts(.optim_output, interactive = FALSE, multiple_keys = FALSE)
}
\arguments{
\item{.optim_output}{optimization output.}

\item{interactive}{logical: Whether or not to return interative plotly graph}

\item{multiple_keys}{Logical. The data has or not multiple keys to plot as grid.}
}
\value{
graph
}
\description{
Plot time series forecast
}
\examples{
\dontrun{
plot_ts()
}
}
