% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_optimization.R
\name{predict_ts}
\alias{predict_ts}
\title{Predicted values from a time series model}
\usage{
predict_ts(
  fit,
  .data,
  ts_model,
  parameter,
  optim_conf,
  add_fitted = logical(),
  type = "response"
)
}
\arguments{
\item{fit}{fit object}

\item{.data}{tibble}

\item{ts_model}{string}

\item{parameter}{list}

\item{optim_conf}{list}

\item{add_fitted}{logical}

\item{type}{string: type of prediction}
}
\value{

}
\description{
Predicted values from a time series model
}
