% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prescribe_ts}
\alias{prescribe_ts}
\title{Data prescription}
\usage{
prescribe_ts(
  .data_init,
  key = NULL,
  y_var,
  date_var,
  reg_name = NULL,
  reg_value = NULL,
  freq,
  date_format
)
}
\arguments{
\item{.data_init}{DataFrame or tibble.}

\item{key}{Column name of the key if any.}

\item{y_var}{Column name of the variable to be forecasted.}

\item{date_var}{Column name of time index.}

\item{reg_name}{Column name of the regressors.}

\item{reg_value}{Column name of the regressors' values.}

\item{freq}{Frequency of the data.}

\item{date_format}{string: date format such as "ymd".}
}
\value{
data-frame
}
\description{
Prescribe provides a simple way to prepare the data for ingestion
}
\examples{
\dontrun{
prescribe_ts()
}
}
