% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{react_gam_par}
\alias{react_gam_par}
\title{Internal GAM parameter}
\usage{
react_gam_par(parameter, values_list)
}
\arguments{
\item{parameter}{list: parameters and hyperparameters.}

\item{values_list}{list: parameter modifiers.}
}
\value{

}
\description{
Internal GAM parameter
}
