% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{reverse_scope}
\alias{reverse_scope}
\title{Reverse scope}
\usage{
reverse_scope(countries, gbus, date_cycle, db = NULL, read_db = FALSE)
}
\arguments{
\item{countries}{string: country codes.}

\item{gbus}{string}

\item{date_cycle}{string: cycle date format yyyy-mm-dd}

\item{db}{string}

\item{read_db}{logical}
}
\value{

}
\description{
This function collects paths or RDatas given countries and GBU
}
