% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{seasonal_features}
\alias{seasonal_features}
\title{Generate trend and seasonal components for regression based models}
\usage{
seasonal_features(
  .data,
  freq = .log$prescription$freq,
  numeric_seas = FALSE,
  hierarchy_seas = FALSE
)
}
\arguments{
\item{.data}{data-frame or tibble}

\item{freq}{Numeric. Time series frequency}

\item{numeric_seas}{logical: whether or not to print numberic seasonal variables instead of factor.}

\item{hierarchy_seas}{logical: whether or not to provide higher aggregation levels. For instance, 
a time series of weekly frequency will generate besides week_seas, also month_seas, and year_seas.}
}
\value{
data-frame or tibble
}
\description{
Generate trend and seasonal components for regression based models
}
\examples{
\dontrun{
get_design_matrix()
}
}
