% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_funs.R
\name{ts_cut}
\alias{ts_cut}
\title{Cut time series}
\usage{
ts_cut(y_var, frequency)
}
\arguments{
\item{y_var}{ts object}

\item{frequency.}{Numeric. Time series frequency.}
}
\value{

}
\description{
This function provides a solution to find breaks in the signal and suggest a new starting point
}
