% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_funs.R
\name{ts_split}
\alias{ts_split}
\title{Automatic Time Series Cross-Validation split}
\usage{
ts_split(data, test_size, lag)
}
\arguments{
\item{data}{DataFrame, tibble or tsibble structures.}

\item{test_size}{Numeric. How many periods will be use to asses the forecast accuracy.}

\item{lag}{Numeric. How many periods ahead to start the test size.}
}
\value{

}
\description{
Automatic Time Series Cross-Validation split
}
