% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{ts_split_all}
\alias{ts_split_all}
\title{Time Series Cross-Validation split}
\usage{
ts_split_all(data, test_size, lag, i)
}
\arguments{
\item{data}{DataFrame or tibble}

\item{test_size}{Numeric. How many periods will be use to asses the forecast accuracy.}

\item{lag}{Numeric. How many periods ahead to start the test size.}

\item{i}{Numeric. Iteration.}
}
\value{

}
\description{
Time series CV split heuristics should keep temporal dependencies, henceforth, the sample should not 
be "shuffled" into train and test. This function uses a different strategy to define the test and train sets
maintain the order of the data.
}
