% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{tscut}
\alias{tscut}
\title{Filter out time series}
\usage{
tscut(y_var, freq = 12)
}
\arguments{
\item{y_var}{Numeric or ts.}

\item{freq}{Time series frequency.}
}
\value{

}
\description{
Filter out time series
}
