% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_ts}
\alias{validate_ts}
\title{Validate Time Series Data}
\usage{
validate_ts(.data, na_values = list(y_var = 0, reg_value = 0, reg_name = ""))
}
\arguments{
\item{.data}{Tibble}

\item{na_values}{list: which values to consider and replace, by default: 
na_values = list(y_var = 0, reg_value = 0, reg_name = "")}
}
\value{
Tibble
}
\description{
This functions fill gaps in time series and provide revelant statistics
}
