% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleansing.R
\name{winsorize_ts}
\alias{winsorize_ts}
\title{Winsorize imputation}
\usage{
winsorize_ts(
  .data,
  freq = 12,
  print_all = FALSE,
  threshold = 0.05,
  impute = FALSE
)
}
\arguments{
\item{freq}{Numeric. Time series frequency, 12 by default.}

\item{print_all}{Logical. Whether to print all time series components.}

\item{y_var}{Numeric. Time series to be cleansed.}

\item{na_marker}{logical or binary indicator that defines which observation should be take out of the
loess decomposition.}
}
\value{
numeric cleansed series
}
\description{
This function replaces outliers by an expected value given the 5% and 95% percentiles of the
error component plus the denoised series.
}
\examples{
\dontrun{
winsorize_ts(AirPassengers)
}
}
\author{
Obryan Poyser
}
