% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoforecast.R
\name{autoforecast}
\alias{autoforecast}
\title{Autoforecast}
\usage{
autoforecast(
  .data,
  parameter,
  test_size = 6,
  lag = 3,
  horizon = 36,
  model,
  optim_profile,
  meta_data = FALSE,
  tune_parallel = FALSE,
  number_best_models = 3,
  pred_interval = FALSE,
  metric = "mape",
  method = "winsorize",
  frequency = 12,
  ...
)
}
\arguments{
\item{.data}{data-frame or tibble}

\item{parameter}{List. Optional set of parameters to estimate models.}

\item{test_size}{Numeric. Number of splits to define the best hyperparameters/model.}

\item{lag}{Numeric. Number of periods ahead to assess model accuracy.}

\item{horizon}{Numeric. Number of periods in the future to generate the forecast.}

\item{model}{String. Name of the model to be estimated.}

\item{optim_profile}{String. Defines how strict show the model look for the best parameter. Options are:
fast, light, medium, and, complete.}

\item{meta_data}{Logical. Whether to return the ranking of models in a list.}

\item{tune_parallel}{Logical. Perform parallelization across different model selection (**experimental**).}

\item{number_best_models}{Integer. Among the best models, how many to output (also controls the Ensemble forecast).}

\item{pred_interval}{Logical. Control if prediction intervals are printed.}

\item{...}{Other parameter from the sub-functions.}
}
\value{
data-frame
}
\description{
This function unifies all modules to bring an automatic forecast given a set of specification.
}
\examples{
\dontrun{
autoforecast()
}
}
