% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoforecast.R
\name{autoforecast}
\alias{autoforecast}
\title{Autoforecast function}
\usage{
autoforecast(
  data,
  frequency = 12,
  models,
  algo_study = TRUE,
  test_size = 6,
  lag = 4,
  output_models = c(1:3),
  clean_series = TRUE,
  parameters = NULL,
  allow_seas = TRUE,
  plot_output = TRUE,
  h = 36
)
}
\arguments{
\item{data}{A data frame with a unique key, y (quantity), date and regressors}

\item{frequency}{By default: 12. Other frequencies will be allowed in next versions}

\item{models}{By default: "All". List of models to apply.}

\item{algo_study}{By default: TRUE. The function will automatically select the best models from the list of models of input and forecast. If it’s in FALSE, the function will just forecast with the selected models.}

\item{test_size}{By default: 6. It’s for the Algo study configuration}

\item{lag}{By default: 4. It’s for the Algo study configuration}

\item{output_models}{By default: 1. How many models for forecasting. You can select which models you want to output: Ex: (1,2,3) = Three best models.}

\item{clean_series}{By default: TRUE. This will perform Kalman’s method where there is an attached regressor. (More methods in progress). This option will also remove 0’s before the time series.}

\item{parameters}{By default: All options in NULL. You can select different parameters for some algorithms}

\item{allow_seas}{By default: TRUE. This will allow seasonal models to be explored and forecasted.}

\item{plot_output}{By default: TRUE. This will save the plots in memory.}

\item{h}{By default: 36. Forecasting horizon}
}
\description{
Autoforecast function
}
