% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoforecast.R
\name{autoforecast}
\alias{autoforecast}
\title{Autoforecast}
\usage{
autoforecast(.data, parameter, test_size, lag, horizon, model, optim_profile)
}
\arguments{
\item{.data}{data-frame or tibble}

\item{parameter}{List. Optional set of parameters to estimate models.}

\item{test_size}{Numeric. Number of splits to define the best hyperparameters/model.}

\item{lag}{Numeric. Number of periods ahead to assess model accuracy.}

\item{horizon}{Numeric. Number of periods in the future to generate the forecast.}

\item{model}{String. Name of the model to be estimated.}

\item{optim_profile}{String. Defines how strict show the model look for the best parameter. Options are:
fast, light, medium, and, complete.}
}
\value{
data-frame
}
\description{
This function unifies all modules to bring an automatic forecast given a set of specification.
}
\examples{
\dontrun{
autoforecast()
}
}
