% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleansing.R
\name{clean_ts}
\alias{clean_ts}
\title{Time Series Cleansing}
\usage{
clean_ts(
  .data,
  freq = numeric(),
  winsorize_config = list(),
  imputation_config = list()
)
}
\arguments{
\item{.data}{DataFrame or tibble.}

\item{freq}{numeric, sets the frequency of the data.}

\item{winsorize_config}{list: options for winsorize cleansing method.}

\item{imputation_config}{list: options for imputation methods. The names can take the 
following configurations: impute_method = {kalman, median, mean, nearest, interpolation}
add_transformation = logical, na_regressor = logical (replaces response variable with NA's and imputates)
, na_missing_dates = logical (replaces response variable with NA's and imputates when missing dates)
, na_value = character (replace NA's by a specific value)}
}
\value{
data-frame, tibble or tsibble.
}
\description{
This function collects the data and applies imputation, time series cuts and leading zeros filtering
}
\examples{
\dontrun{
clean_ts()
}
}
