% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleansing.R
\name{clean_ts}
\alias{clean_ts}
\title{Time Series Cleansing}
\usage{
clean_ts(
  .data,
  method = "kalman",
  na_exclude = NULL,
  replace_y_var = TRUE,
  print_all = FALSE,
  threshold = 0.05,
  impute = FALSE,
  winsorize = TRUE
)
}
\arguments{
\item{.data}{DataFrame or tibble.}

\item{method}{Method to be applied to the time series. Options: kalman, winsorize, 
mean, median, nearest observation, interpolation. If more than one method is chosen, a set of new columns
is automatically generated.}

\item{na_exclude}{Vector of names that **should not** be used to replace the observation by NA.}

\item{replace_y_var}{Logical. If replace = TRUE (default) it will 
replace the previous y_var by its cleansed/imputated version. Otherwise, a "y_var_clean" column will
be attached to the data matrix.}

\item{y_var}{String. Colname of the variable to be cleansed.}

\item{date_var}{String. Colname of time serie index.}

\item{freq}{Time series frequency.}
}
\value{
data-frame, tibble or tsibble.
}
\description{
This function collects the data and applies imputation, time series cuts and leading zeros filtering
}
\examples{
\dontrun{
clean_ts()
}
}
