% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_engineering_ts}
\alias{feature_engineering_ts}
\title{Automatic Time Series Feature Engineering}
\usage{
feature_engineering_ts(
  .data,
  lag_var = list(),
  ma_var = list(),
  numeric_seas = FALSE,
  hierarchy_seas = FALSE
)
}
\arguments{
\item{.data}{data-frame or tibble}

\item{lag_var}{list: defines the number of lag periods for a given variable, for instance, 
list(y_var = 1) will generate a lag (autoregressive) variable of 1 period.}

\item{ma_var}{list: defines the number of periods before and after to apply a mean, for instance, 
list(y_var = c(1, 2)) will generate a moving average of 1 period before and 2 after a given point.}

\item{numeric_seas}{logical: whether or not to print numberic seasonal variables instead of factor.}

\item{hierarchy_seas}{logical: whether or not to provide higher aggregation levels. For instance, 
a time series of weekly frequency will generate besides week_seas, also month_seas, and year_seas.}
}
\value{
data-frame or tibble
}
\description{
This function applies different heuristics to add time series features to the original data.
}
\examples{
\dontrun{
feature_engineering_ts()
}
}
