% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_engineering_ts}
\alias{feature_engineering_ts}
\title{Automatic Time Series Feature Engineering}
\usage{
feature_engineering_ts(.data)
}
\arguments{
\item{.data}{data-frame or tibble}
}
\value{
data-frame or tibble
}
\description{
This function applies different heuristics to add time series features to the original data.
}
\examples{
\dontrun{
feature_engineering_ts()
}
}
