% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ets.R
\name{fit_ets}
\alias{fit_ets}
\title{Fit Error Trend Seasonal model}
\usage{
fit_ets(.data, parameter = NULL)
}
\arguments{
\item{.data}{Data frame or tibble.}

\item{parameter}{List. Combination of parameter to estimate the model.}

\item{y_var}{String. Column name of the time series to be forecasted.}
}
\value{
data-frame
}
\description{
Fit Error Trend Seasonal model
}
\examples{
\dontrun{
get_ets()
}
}
