% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_optimization.R
\name{fit_ts}
\alias{fit_ts}
\title{Fitting models for time-series data}
\usage{
fit_ts(.data, ts_model, parameter = list())
}
\arguments{
\item{.data}{tibble/data.frame: matrix of response and covariates.}

\item{ts_model}{string: name of the mode to be fitted.}

\item{parameter}{list: Combination of parameter to estimate the model.}
}
\value{
data-frame
}
\description{
Wrapper of single function call to simplify the estimation
}
\examples{
\dontrun{
fit_ts()
}
}
