% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ets.R
\name{get_ets}
\alias{get_ets}
\title{Fit Error Trend Seasonal model}
\usage{
get_ets(.data, y_var, parameter = NULL)
}
\arguments{
\item{.data}{Data frame or tibble.}

\item{y_var}{String. Column name of the time series to be forecasted.}

\item{parameter}{List. Combination of parameter to estimate the model.}
}
\value{
data-frame
}
\description{
Fit Error Trend Seasonal model
}
\examples{
\dontrun{
get_ets()
}
}
