% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neural_network.R
\name{get_neural_network}
\alias{get_neural_network}
\title{Fit Neural Network model}
\usage{
get_neural_network(.data, y_var, x_data = NULL, parameter = NULL)
}
\arguments{
\item{.data}{Data frame or tibble with a response variable.}

\item{y_var}{String. Column name of the time series to be forecasted.}

\item{x_data}{Data frame. Design matrix to calculate predicted/forecast figures.}

\item{parameter}{List. Parameter to be used for estimation.}
}
\value{
data-frame
}
\description{
Fit Neural Network model
}
\examples{
\dontrun{
get_neural_network()
}
}
