% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seasonal_naive.R
\name{get_seasonal_naive}
\alias{get_seasonal_naive}
\title{Fit a Seasonal Naive Model}
\usage{
get_seasonal_naive(.data, y_var, parameter = NULL, horizon)
}
\arguments{
\item{.data}{Data frame or tibble with a response variable.}

\item{y_var}{String. Column name of the time series to be forecasted.}

\item{parameter}{List. Optional parameters.}

\item{horizon}{Numeric. Number of periods ahead to forecast.}
}
\value{
data-frame or tibble
}
\description{
Fit a Seasonal Naive Model
}
\examples{
\dontrun{
get_seasonal_naive()
}
}
