% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_funs.R
\name{get_time_weights}
\alias{get_time_weights}
\title{Generate time weights}
\usage{
get_time_weights(dep_var, time_weight)
}
\arguments{
\item{dep_var}{numeric. Time series vector data.}

\item{time_weight}{numeric. How rapidly recent observations weight for estimation.}
}
\value{

}
\description{
Generate time weights
}
\author{
Obryan Poyser
}
