% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_funs.R
\name{get_trend_discounts}
\alias{get_trend_discounts}
\title{Generate trend discounts}
\usage{
get_trend_discounts(y_var_length, trend_discount, horizon = NULL, lag = NULL)
}
\arguments{
\item{y_var_length}{Numeric. Length of the time series}

\item{trend_discount}{Numeric. How rapidly the trend reach the stability}

\item{horizon}{Numeric. How far in time to produce trend discounts.}

\item{lag}{Numeric. Lag to be used for cross-validation purposes.}
}
\value{

}
\description{
Regression based models use a linear trend to account for the change in level over time. 
In practical terms, it is measured as a vector of equidistant integers.
Often, the trend component can significantly impact the forecast in the long run. One way to solve
this issue is to apply a "discount" on the trend vector, henceforth, reducing the marginal effect on the 
predictions.
}
\author{
Obryan Poyser
}
