% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansing.R
\name{impute_ts}
\alias{impute_ts}
\title{Time series imputation}
\usage{
impute_ts(
  .data,
  y_var,
  method = "winsorize",
  na_exclude = NULL,
  frequency = 12,
  replace = TRUE,
  ...
)
}
\arguments{
\item{.data}{DataFrame or tibble.}

\item{y_var}{Quoted or unquoted variable name to be cleansed.}

\item{method}{Method to be applied to the time series. Options: kalman, winsorize, 
mean, median, nearest observation, interpolation. If more than one method is chosen, a set of new columns
is automatically generated.}

\item{na_exclude}{Vector of names that **should not** be used to replace the observation by NA.}

\item{frequency}{Time series frequency.}

\item{replace}{Logical. If replace = TRUE (default) it will 
replace the previous y_var by its cleansed/imputated version. Otherwise, a "y_clean" column will
be attached to the data matrix.}

\item{...}{Other parameter}
}
\value{

}
\description{
Time series imputation
}
\author{
Obryan Poyser
}
