% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_funs.R
\name{make_reg_matrix}
\alias{make_reg_matrix}
\title{Generate regression matrix for regression based models}
\usage{
make_reg_matrix(fit_output, horizon)
}
\arguments{
\item{fit_output}{metadata from regression fit}

\item{horizon}{numeric. How far in time to produce a xvar matrix.}
}
\value{

}
\description{
This function takes the metadata from the fit to generate a template of the
dependent variables to make forecast.
}
