% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{mape}
\alias{mape}
\title{MAPE}
\usage{
mape(real, pred)
}
\arguments{
\item{real}{Numeric. Observed value.}

\item{pred}{Numeric. Predicted value.}
}
\value{

}
\description{
Mean Average Prediction Error forecast accuracy metric
}
