% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansing.R
\name{na_winsorize}
\alias{na_winsorize}
\title{Winsorize imputation}
\usage{
na_winsorize(series, na_marker = NULL, frequency = 12, print_all = FALSE)
}
\arguments{
\item{series}{numeric or ts. Time series to be cleansed.}

\item{na_marker}{logical or binary indicator that defines which observation should be take out of the
loess decomposition.}

\item{frequency}{numeric. Time series frequency}

\item{print_all}{logical. Whether to print all time series components.}
}
\value{
numeric cleansed series
}
\description{
This function replaces outliers by an expected value given the 5% and 95% percentiles of the
error component plus the denoised series.
}
\examples{
\dontrun{
na_winsorize(AirPassengers)
}
}
\author{
Obryan Poyser
}
