% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_ts.R
\name{optim_ts}
\alias{optim_ts}
\title{Hyperparameter optimization for a glmnet model}
\usage{
optim_ts(.data, test_size, lag, parameter, model, tune_parallel = FALSE)
}
\arguments{
\item{.data}{Data frame or tibble.}

\item{test_size}{Numeric. How many periods will be use to asses the forecast accuracy.}

\item{lag}{Numeric. How many periods ahead to start the test size.}

\item{parameter}{List. Hyperparameters to be used for estimation. There are 4 hyperparameters: first, 
*alpha* in the space [0,1] controls whether it is a Ridge (L2) a LASSO (L1) shrinkage method, respectively.
Any number that lies between is considered as ElasticNet regression, a combination of both regularizations.
The other 2 hyperparameters are time weights and trend discount.}

\item{model}{String. Model to be optimized.}

\item{tune_parallel}{Logical. Perform parallelization across different model selection (**experimental**).}
}
\value{
data-frame or tibble
}
\description{
Hyperparameter optimization for a glmnet model
}
\examples{
\dontrun{
optim_ts()
}
}
