% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoforecast.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot time series forecast}
\usage{
plot_ts(.optim_output, interactive = FALSE, multiple_keys = FALSE)
}
\arguments{
\item{interactive}{Logical. Whether or not to return interative plotly graph}

\item{multiple_keys}{Logical. The data has or not multiple keys to plot as grid.}

\item{.data}{data-frame of class optim_output}
}
\value{
graph
}
\description{
Plot time series forecast
}
\examples{
\dontrun{
plot_ts()
}
}
