% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansing.R
\name{prescribe_ts}
\alias{prescribe_ts}
\title{Data prescription}
\usage{
prescribe_ts(.data, key, y_var, date_var, freq, extend_design = TRUE)
}
\arguments{
\item{.data}{DataFrame or tibble.}

\item{key}{Column name of the key if any.}

\item{y_var}{Column name of the variable to be forecasted.}

\item{date_var}{Column name of time index.}

\item{freq}{Frequency of the data.}

\item{extend_design}{Logical. Print the data, TRUE by default.}
}
\value{

}
\description{
Prescribe provides a simple way to define the metadata
}
