% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{split_ts}
\alias{split_ts}
\title{Automatic Time Series Cross-Validation split}
\usage{
split_ts(.data, test_size, lag)
}
\arguments{
\item{.data}{DataFrame, tibble or tsibble structures.}

\item{test_size}{Numeric. How many periods will be use to asses the forecast accuracy.}

\item{lag}{Numeric. How many periods ahead to start the test size.}
}
\value{

}
\description{
Automatic Time Series Cross-Validation split
}
