% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{update_parameter}
\alias{update_parameter}
\title{Updating parameters given a grid}
\usage{
update_parameter(parameter, new_parameter, model = "glmnet")
}
\arguments{
\item{parameter}{List. Parameters to be used to estimate model or define the job to be done.}

\item{new_parameter}{Named list of data frame.}

\item{model}{String. Model where the parameters will be replaced.}
}
\value{

}
\description{
Updating parameters given a grid
}
