% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_ts}
\alias{validate_ts}
\title{Validate Time Series Data}
\usage{
validate_ts(.data, na_value = 0)
}
\arguments{
\item{.data}{Tibble}
}
\value{
Tibble
}
\description{
This functions fill gaps in time series and provide revelant statistics
}
