% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleansing.R
\name{winsorize_ts}
\alias{winsorize_ts}
\title{Winsorize imputation}
\usage{
winsorize_ts(.data, freq = numeric(), winsorize_config = list())
}
\arguments{
\item{.data}{tibble/data.frame: data.}

\item{freq}{integer: frequency of the time series.}

\item{winsorize_config}{list: options for winsorize cleansing method.}
}
\value{
numeric cleansed series
}
\description{
This function replaces outliers by an expected value given the 5% and 95% percentiles of the
error component plus the denoised series.
}
\examples{
\dontrun{
winsorize_ts(AirPassengers)
}
}
\author{
Obryan Poyser
}
